requirements.txt
schedule==1.2.2
holidays
apscheduler
python-dotenv
websockets
nest-asyncio==1.6.0
slack_sdk